+++ libs/rxjs/utility/src/lib/weight.spec.ts
@@ -16,7 +16,7 @@ describe('weight', () => {
         b: 100,
         c: 1000,
       });
-      m.expect(input.pipe(weight('foobar' as any, Weights.KG))).toBeObservable(expected);
+      m.expect(input.pipe(weight('foobar' as any, Weights.KILOGRAMS))).toBeObservable(expected);
       m.expect(input).toHaveSubscriptions(subs);
     }),
   );
@@ -37,7 +37,7 @@ describe('weight', () => {
           b: 0.1,
           c: 1,
         });
-        m.expect(input.pipe(weight(Weights.G, Weights.KG))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.GRAMS, Weights.KILOGRAMS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -56,7 +56,7 @@ describe('weight', () => {
           b: 0.22,
           c: 2.205,
         });
-        m.expect(input.pipe(weight(Weights.G, Weights.LB))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.GRAMS, Weights.POUNDS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -75,7 +75,7 @@ describe('weight', () => {
           b: 3.527,
           c: 35.274,
         });
-        m.expect(input.pipe(weight(Weights.G, Weights.OZ))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.GRAMS, Weights.OUNCES))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -94,7 +94,7 @@ describe('weight', () => {
           b: 0.016,
           c: 0.157,
         });
-        m.expect(input.pipe(weight(Weights.G, Weights.ST))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.GRAMS, Weights.STONE))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -116,7 +116,7 @@ describe('weight', () => {
           b: 100000,
           c: 1000000,
         });
-        m.expect(input.pipe(weight(Weights.KG, Weights.G))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.KILOGRAMS, Weights.GRAMS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -135,7 +135,7 @@ describe('weight', () => {
           b: 220.46,
           c: 2204.6,
         });
-        m.expect(input.pipe(weight(Weights.KG, Weights.LB))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.KILOGRAMS, Weights.POUNDS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -154,7 +154,7 @@ describe('weight', () => {
           b: 3527.4,
           c: 35274,
         });
-        m.expect(input.pipe(weight(Weights.KG, Weights.OZ))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.KILOGRAMS, Weights.OUNCES))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -173,7 +173,7 @@ describe('weight', () => {
           b: 15.74,
           c: 157.4,
         });
-        m.expect(input.pipe(weight(Weights.KG, Weights.ST))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.KILOGRAMS, Weights.STONE))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -195,7 +195,7 @@ describe('weight', () => {
           b: 45359.702,
           c: 453597.024,
         });
-        m.expect(input.pipe(weight(Weights.LB, Weights.G))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.POUNDS, Weights.GRAMS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -214,7 +214,7 @@ describe('weight', () => {
           b: 45.36,
           c: 453.597,
         });
-        m.expect(input.pipe(weight(Weights.LB, Weights.KG))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.POUNDS, Weights.KILOGRAMS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -233,7 +233,7 @@ describe('weight', () => {
           b: 1600,
           c: 16000,
         });
-        m.expect(input.pipe(weight(Weights.LB, Weights.OZ))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.POUNDS, Weights.OUNCES))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -252,7 +252,7 @@ describe('weight', () => {
           b: 7.143,
           c: 71.429,
         });
-        m.expect(input.pipe(weight(Weights.LB, Weights.ST))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.POUNDS, Weights.STONE))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -274,7 +274,7 @@ describe('weight', () => {
           b: 2834.949,
           c: 28349.493,
         });
-        m.expect(input.pipe(weight(Weights.OZ, Weights.G))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.OUNCES, Weights.GRAMS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -293,7 +293,7 @@ describe('weight', () => {
           b: 2.835,
           c: 28.349,
         });
-        m.expect(input.pipe(weight(Weights.OZ, Weights.KG))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.OUNCES, Weights.KILOGRAMS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -312,7 +312,7 @@ describe('weight', () => {
           b: 6.25,
           c: 62.5,
         });
-        m.expect(input.pipe(weight(Weights.OZ, Weights.LB))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.OUNCES, Weights.POUNDS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -331,7 +331,7 @@ describe('weight', () => {
           b: 0.446,
           c: 4.464,
         });
-        m.expect(input.pipe(weight(Weights.OZ, Weights.ST))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.OUNCES, Weights.STONE))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -353,7 +353,7 @@ describe('weight', () => {
           b: 635041.595,
           c: 6350415.952,
         });
-        m.expect(input.pipe(weight(Weights.ST, Weights.G))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.STONE, Weights.GRAMS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -372,7 +372,7 @@ describe('weight', () => {
           b: 635.042,
           c: 6350.416,
         });
-        m.expect(input.pipe(weight(Weights.ST, Weights.KG))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.STONE, Weights.KILOGRAMS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -391,7 +391,7 @@ describe('weight', () => {
           b: 1400,
           c: 14000,
         });
-        m.expect(input.pipe(weight(Weights.ST, Weights.LB))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.STONE, Weights.POUNDS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -410,7 +410,7 @@ describe('weight', () => {
           b: 22400,
           c: 224000,
         });
-        m.expect(input.pipe(weight(Weights.ST, Weights.OZ))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.STONE, Weights.OUNCES))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );